rag-project/requirements.txt
flask
groq
sentence-transformers
faiss-cpu
pypdf
python-dotenv
